Bytter 'paradigme' for å kombinere tidslinjer fra LocalDataTimeline-baserte til 'snitt-baserte' fordi førstnevnte av og til tilsynelatendee sliter med korte perioder og lager overlapp
